+++ .gitignore
@@ -13,7 +13,7 @@ __pycache__/
 
 # Distribution / packaging
 .Python
-build/
+/build
 develop-eggs/
 dist/
 downloads/
