+++ powerflow/decorators/__init__.py
@@ -1,4 +1,14 @@
-def pf_tools(func):
+from functools import wraps
+
+
+def pf_tools(func: callable):
+    """[summary]
+
+    :param func: [description]
+    :type func: callable
+    """
+
+    @wraps(func)
     def wrapper(distSys, *args, **kwds):
         if distSys.dss.Circuit.Name() != distSys.get_name():
             distSys.init_sys()
