+++ pyproject.toml
@@ -14,6 +14,7 @@ flake8 = "^3.7.9"
 black = "^19.10b0"
 ipython = "^7.13.0"
 ipykernel = "^5.2.1"
+guzzle_sphinx_theme = "^0.7.11"
 
 [build-system]
 requires = ["poetry>=0.12"]
